Implement new design and close details
